fix: map cannot have empty line #63
